Properly round consolidator updates
This check ensures that the data is on the subscribed interval.
A bug was found where daily data with a different data tz and
exchange tz was not being pumped into the consolidator. This is
because we were rounding a time such as 20:00 by 1-day. If we
convert the time to the data time zone, it becomes 00:00 and
then the rounding operation returns the same input value which
signals that we should use this data point to update consolidtors.